Los campos de la configuración de dimensiones para expediciones en la ficha del producto y la tabla de dimensiones y pesos de los albaranes, tienen que estar en número real, no entero.

+++ project-addons/custom_views/models/product.py
@@ -261,9 +261,9 @@ class ProductTemplate(models.Model):
                                      inverse_name='product_id')
     ecoembes_weight = fields.Float(digits = dp.get_precision('Stock Weight'))
     expeditions_name = fields.Char('Expeditions name')
-    expeditions_width = fields.Integer('Width (cm)')
-    expeditions_height = fields.Integer('Height (cm)')
-    expeditions_depth = fields.Integer('Depth (cm)')
+    expeditions_width = fields.Float('Width (cm)')
+    expeditions_height = fields.Float('Height (cm)')
+    expeditions_depth = fields.Float('Depth (cm)')
 
     @api.one
     @api.depends('seller_ids')
